chore: move and clear the TODO for GeckoLib
